requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
umap-learn>=0.5.1
imageio>=2.9.0
plotly>=5.3.0
kaleido>=0.2.1
scipy>=1.7.0
Pillow>=8.2.0
openpyxl>=3.0.7
tqdm>=4.61.0 